+++ src/index.ts
@@ -10,6 +10,7 @@ export * from "./lib/DatePickerInput";
 export * from "./lib/ColorPickerInput";
 export * from "./lib/RatingInput";
 export * from "./lib/helpers/dateUtils";
+export * from "./lib/helpers/form";
 export * from "./lib/helpers/mui";
 export * from "./lib/helpers/typeahead";
 export * from "./lib/hooks/useDebounceHook";