+++ CHANGELOG.md
@@ -7,6 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Required field label on `FormGroupLayoutLabel`, `ColorPicker`, `TelephoneNumberInput`, `TypeaheadTextField` (hence `StaticTypeaheadInput` and `AsyncTypeaheadInput`) in order to display \* also on nested and array fields.
+
+  1. `requiredFields` still needs `FieldPath<T>[]`
+  2. Arrays need to provide an index notation in order to be complaint with `FieldPath` react-hook-form type (`object.${number}.property`). Example:
+
+    ```tsx
+    requiredFields = [`object`, `objects.0.property`, `object.nestedObject.property`, `object.nestedObjects.1.property`];
+    ```
+
+    is going to consider as required:
+
+    ```tsx
+      name="object"
+      name="objects.0.property", name="objects.1.property", etc.
+      name="object.nestedObject.property"
+      name="object.nestedObjects.0.property", name="object.nestedObjects.1.property", etc.
+    ```
+
+### Added
+
+- `form` helper functions
+
 ## [3.13.1] - 2025-11-13
 
 ### Fixed